fixed frontend display
